fixed an overflow on large maps, added comments, made the Makefile more ugly

+++ globals.cpp
@@ -1,8 +1,7 @@
 #include "globals.h"
-#include <cstdlib>
 
 int S_FROMX = UNDEFINED, S_FROMZ = UNDEFINED, S_TOX = UNDEFINED, S_TOZ = UNDEFINED;
-int MAPSIZE_Y = 128, MAPSIZE_Z = 0, MAPSIZE_X = 0;
+size_t MAPSIZE_Y = 128, MAPSIZE_Z = 0, MAPSIZE_X = 0;
 bool g_Nightmode = false;
 bool g_Underground = false;
 bool g_Skylight = false;
